+++ apps/backoffice-v2/src/Router/Router.tsx
@@ -21,6 +21,9 @@ import { NotFoundRedirect } from '@/pages/NotFound/NotFound';
 import { TransactionMonitoringAlerts } from '@/pages/TransactionMonitoringAlerts/TransactionMonitoringAlerts.page';
 import { TransactionMonitoring } from '@/pages/TransactionMonitoring/TransactionMonitoring';
 import { TransactionMonitoringAlertsAnalysisPage } from '@/pages/TransactionMonitoringAlertsAnalysis/TransactionMonitoringAlertsAnalysis.page';
+import { Home } from '@/pages/Home/Home.page';
+import { Statistics } from '@/pages/Statistics/Statistics.page';
+import { Workflows } from '@/pages/Workflows/Workflows.page';
 import { Profiles } from '@/pages/Profiles/Profiles.page';
 import { Individuals } from '@/pages/Profiles/Individuals/Individuals.page';
 import { MerchantMonitoring } from '@/pages/MerchantMonitoring/MerchantMonitoring.page';
@@ -145,6 +148,23 @@ const router = createBrowserRouter([
                     ],
                   },
                 ],
+              },
+              {
+                path: '/:locale/home',
+                element: <Home />,
+                children: [
+                  {
+                    path: '/:locale/home/statistics',
+                    element: <Statistics />,
+                    errorElement: <RouteError />,
+                  },
+                  {
+                    path: '/:locale/home/workflows',
+                    element: <Workflows />,
+                    errorElement: <RouteError />,
+                  },
+                ],
+                errorElement: <RouteError />,
               },
             ],
           },